Refactor: 디렉토리 구조에 맞춰 코드 이동
