--- fastapi/Makefile
@@ -1,2 +1,0 @@
-all:
-	uvicorn app:app --host 0.0.0.0 --port 3000 --reload
